Split web repo routers to different directories
