+++ routers/web/repo/base/alert.go
@@ -1,0 +1,8 @@
+// Copyright 2024 The Gitea Authors. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+package base
+
+import "code.gitea.io/gitea/modules/base"
+
+var TplAlertDetails base.TplName = "base/alert_details"
