+++ routers/web/repo/issue/issue_content_history.go
@@ -1,7 +1,7 @@
 // Copyright 2021 The Gitea Authors. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-package repo
+package issue
 
 import (
 	"bytes"
