+++ routers/web/repo/issue/issue_timetrack.go
@@ -1,7 +1,7 @@
 // Copyright 2017 The Gitea Authors. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-package repo
+package issue
 
 import (
 	"net/http"
